refactor(palette-tauri): simplify after /simplify review
- Extract validate_saved_server_url to lib.rs (was duplicated verbatim in
  axon_bridge.rs and stream.rs — removes divergence risk)
- Add StreamClient to AppState so axon_http_stream_request shares a pooled
  connection instead of building a new reqwest::Client per SSE request
- Move axon_bridge inline mod tests to sidecar axon_bridge_tests.rs to
  match the project-wide #[path] convention
- Simplify shortcut merge: drop no-op .or(Some(...)) before .unwrap_or_else
- Clarify double-negation in persistence.rs: use map_or instead of is_none_or
